% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_funcs.R
\name{fav_by_author}
\alias{fav_by_author}
\title{retrieve my twitter favorites from a given user}
\usage{
fav_by_author(likes_user = "stchlk", twitter_user, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
retrieve my twitter favorites from a given user
}
