% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers.R
\name{transmute_pairwise}
\alias{transmute_pairwise}
\title{Transmute Pairwise}
\usage{
transmute_pairwise(df, fun, ..., associative = TRUE)
}
\arguments{
\item{associative}{}
}
\value{

}
\description{
return a dataframe consisting of all combinations of arbitrary pairwise operations across a selection of columns
}
\examples{
mpg \%>\%
select(where(is.numeric)) \%>\%
 mutate(select(cur_data(), c(displ, cty, hwy)) \%>\%
          transmute_pairwise(`/`, associative = FALSE) \%>\%
          rename_with(~paste0(.x, "_ratio")) \%>\%
          identity()
 )
}
