% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_funcs.R
\name{fav_by_author}
\alias{fav_by_author}
\title{retrieve my twitter favorites from a given user}
\usage{
fav_by_author(twitter_user, likes_user = "stchlk", ...)
}
\arguments{
\item{twitter_user}{the user whose tweets were favorited}

\item{likes_user}{the user whose favorite we'll query}

\item{...}{additional arguments passed to \code{rtweet::get_favorites}}
}
\value{

}
\description{
retrieve my twitter favorites from a given user
}
